backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
feedparser==6.0.11
pydantic>=2.9.0
sqlalchemy==2.0.23
psycopg[binary]>=3.2.2
alembic==1.13.1
python-dotenv==1.0.0
anthropic>=0.76.0
reportlab==4.0.9
Pillow==10.4.0
